Refactor IP restriction middleware with configurable office-only access
